requirements.txt
python-dotenv==1.0.1
requests==2.32.3
Flask==3.1.0
matplotlib==3.8.2
pillow==10.2.0
